requirements.txt
streamlit
google-generativeai
langchain-google-genai
langchain-core
langchain-text-splitters
faiss-cpu
sentence-transformers
langchain-community
Pillow
fpdf2

# Forcing a hard reset v2